Working on stack initialization

+++ src/file_parsing/parse_line.c
@@ -22,7 +22,7 @@ void	parse_line(t_file *file)
 	if (file->current_line == INVALID)
 		__exit(file, E_INVALID_LINE);
 	is_duplicate(file);
-	skip_line_start(file->current_line, &file->line);
+	skip_line_start((int)file->current_line, &file->line);
 	if (!is_line_valid(file->instructions, file))
 		__exit(file, E_INVALID_LINE);
 	set_duplicate_check(file);
